ACC: Support ACC ramping based on CPU Load
* Rearrange acc_ramp struct and API to allow other variables

* Move ACC ramp chan-load VTY cmds to an own node
Since we'll be adding more parameters and naming is already quite long,
let's move everything to an own node instead, to keep names smaller as
well as being able to identify them when we add CPU load checks.
Old commands can still be used but are marked as deprecated in favour of
new commands in the created node.

* Add new VTY nodes to configure ACC ramping based on CPU load.

Related: SYS#4939
Change-Id: I1930f5c3af34e648021ffdcd455e2c72821edbd8

WIP: cpuload

Change-Id: I2a74f5f695237d731db467924c1b6609eadce9a9