RSL: look up lchan by sign_link, not trx, for VAMOS
So far we pass the gsm_bts_trx to the lchan lookup functions to
determine the lchan that a specific RSL message is received for. But for
VAMOS secondary lchans, we introduced a second sign_link for each TRX,
so passing the TRX loses information on which TEI the RSL message was
received from. Pass the sign_link instead of the trx.

If the sign_link is the secondary VAMOS link, return the VAMOS shadow
lchan instead of the primary one for that lchan number.

Related: SYS#5315 OS#4940
Change-Id: I872a3ed9ce0cdbd9fbe41a750081f54008eece8d