RR Assignment for VAMOS: send TSC Set
We already send the TSC Set in the RSL Channel Activation, telling the
BTS about the TSC Set. The MS needs to be instructed of the TSC Set in
the RR Assignment Command.

The first code to actually do an Assignment to a VAMOS activated lchan
will follow in If006f5caaf83b07675f57e5665cfa79328da55e6.

Change-Id: Ibf3b6d276fadf724c16a5225c03e96a27a056153