WIP: Avoid autofilling neigh list if no neigh is configured
This piece of code was still in osmo-bsc in order to keep backward
compatibility with older deployments from the time where configuring
neighbours was not possible.
We have support to configure neighbors since a long time ago now, and
this backward-compatible behavior has more drawbacks than benefits:
* It is totally acceptable for a cell to have no neighbors, so currently
  there's no way to enfore it.
* If ANR is in use, a usual case is that each cell starts with no
  neighbors and then dynamic ones are added after ANR procedure runs.

So let's drop this behavior and avoid autofilling the list if no
neighbor is selected.

Related: SYS#5303
Change-Id: Iff1c7245250e8cca81743f6540d3d317ba09e35c